refactor to be more kubey based on external comments
